changed the size_t to int in kernels

+++ kernel_files/macros_cl.cl
@@ -137,7 +137,7 @@ typedef struct kernel_info_struct {
     #if 0
     #define REDUCTION(in, out, operation)                    \
         barrier(CLK_LOCAL_MEM_FENCE);                               \
-        for (size_t offset = BLOCK_SZ / 2; offset > 0; offset /= 2) \
+        for (int offset = BLOCK_SZ / 2; offset > 0; offset /= 2) \
         {                                                           \
             if (lid < offset)                                       \
             {                                                       \
@@ -154,14 +154,14 @@ typedef struct kernel_info_struct {
     #define REDUCTION(in, out, operation)                    \
     { \
         barrier(CLK_LOCAL_MEM_FENCE);                               \
-        const size_t vecsz = 512/(sizeof(in[0])*8);  \
-        const size_t redsz = vecsz*2; \
-        size_t remain = BLOCK_SZ ; \
+        const int vecsz = 512/(sizeof(in[0])*8);  \
+        const int redsz = vecsz*2; \
+        int remain = BLOCK_SZ ; \
         do \
         { \
             if (!(lid % redsz) && lid < remain)                \
             {                                                           \
-                /*for (size_t offset = 1; offset < redsz; offset++)    \
+                /*for (int offset = 1; offset < redsz; offset++)    \
                 {                                                       \
                     in[lid] = operation(in[lid], in[lid + offset]);               \
                 }*/                                                       \
@@ -173,7 +173,7 @@ typedef struct kernel_info_struct {
                 in[5 + lid] = operation(in[5 + lid], in[5 + lid+vecsz]); \
                 in[6 + lid] = operation(in[6 + lid], in[6 + lid+vecsz]); \
                 in[7 + lid] = operation(in[7 + lid], in[7 + lid+vecsz]);*/ \
-                for (size_t offset = 0; offset < vecsz; offset++)    \
+                for (int offset = 0; offset < vecsz; offset++)    \
                 {                                                       \
                     in[offset + lid] = operation(in[offset + lid], in[offset + lid+vecsz]); \
                 }                                                       \
@@ -189,13 +189,13 @@ typedef struct kernel_info_struct {
         out[get_group_id(1)*get_num_groups(0) + get_group_id(0)] = in[0]; \
         /*if (0 == lid)                                               \
         {                                                           \
-            for (size_t offset = 1; offset < BLOCK_SZ/vecsz; offset++)    \
+            for (int offset = 1; offset < BLOCK_SZ/vecsz; offset++)    \
             {                                                       \
                 in[0] = operation(in[0], in[offset]);               \
             }                                                       \
             out[get_group_id(1)*get_num_groups(0) + get_group_id(0)] = in[0]; \
         }*/ \
-        /*for (size_t offset = BLOCK_SZ / (2*vecsz); offset > 0; offset /= 2) \
+        /*for (int offset = BLOCK_SZ / (2*vecsz); offset > 0; offset /= 2) \
         {                                                           \
             if (lid < offset)                                       \
             {                                                       \
